+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.1
+## 1.0.1.0
+- Added success message when changing gamemode.
+- Added error message when an invalid gamemode is specified.
+
+## 1.0.0.0
 - Added `/gm <gamemode>` for changing gamemode.
 - Added `/gm{0|s|survival}` for changing gamemode to Survival.
 - Added `/gm{1|c|creative}` for changing gamemode to Creative.